+++ tlatools/src/tlc2/tool/fp/FPSetStatistic.java
@@ -12,7 +12,7 @@ public interface FPSetStatistic {
 	/**
 	 * @return The allocated (used and unused) array length of the first level in-memory storage.
 	 */
-	public int getTblCapacity();
+	public long getTblCapacity();
 
 	/**
 	 * @return the index.length
